requirements.txt
Flask==3.1.0
requests==2.32.3
beautifulsoup4==4.12.3

openai==1.21.2
python-dotenv==1.0.1

httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0

langchain==0.1.16
langchain-chroma==0.1.0
langchain-community==0.0.33
langchain-core==0.1.44
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.48
jq
gunicorn
